+++ lib/websocket/connection.js
@@ -2,16 +2,12 @@
 
 const { EventEmitter } = require('node:events');
 
-const {
-  OPCODES,
-  MAX_BUFFER,
-  CLOSE_TIMEOUT,
-  DATA_OPCODES,
-  CLOSE_CODES,
-  PROTOCOL_ERROR_SUBTYPES,
-} = require('./constants.js');
+const { OPCODES, DATA_OPCODES, CLOSE_CODES } = require('./constants.js');
 const { Frame } = require('./frame.js');
 const { FrameParser, isValidUTF8 } = require('./frameParser.js');
+
+const MAX_BUFFER = 1024 * 1024 * 100;
+const CLOSE_TIMEOUT = 1000;
 
 class Connection extends EventEmitter {
   #socket;
@@ -87,16 +83,11 @@ class Connection extends EventEmitter {
 
       const { frame, bytesUsed } = result.value;
       if (!this.#isClient && !frame.masked) {
-        return void this.#close(
-          Frame.protocolErrorClose(PROTOCOL_ERROR_SUBTYPES.UNMASKED),
-        );
+        return void this.#close(Frame.protocolErrorClose('UNMASKED'));
       }
       if (this.#isClient && frame.masked) {
         return void this.#close(
-          Frame.protocolErrorClose(
-            PROTOCOL_ERROR_SUBTYPES.MASKED,
-            this.#isClient,
-          ),
+          Frame.protocolErrorClose('MASKED', this.#isClient),
         );
       }
 
@@ -144,12 +135,7 @@ class Connection extends EventEmitter {
     const [type, subtype] = code.split('-');
     this.emit('error', error);
     if (type === 'PROTOCOL_ERROR') {
-      this.#close(
-        Frame.protocolErrorClose(
-          PROTOCOL_ERROR_SUBTYPES[subtype],
-          this.#isClient,
-        ),
-      );
+      this.#close(Frame.protocolErrorClose(subtype, this.#isClient));
     } else {
       this.#close(Frame.errorClose(type, this.#isClient));
     }
@@ -178,10 +164,7 @@ class Connection extends EventEmitter {
           new Error('Protocol error: Unexpected CONTINUATION without start'),
         );
         return void this.#close(
-          Frame.protocolErrorClose(
-            PROTOCOL_ERROR_SUBTYPES.COMMON,
-            this.#isClient,
-          ),
+          Frame.protocolErrorClose('COMMON', this.#isClient),
         );
       }
       if (frame.fin) {
@@ -219,12 +202,7 @@ class Connection extends EventEmitter {
         'error',
         new Error('Protocol error: Unexpected data frame during fragments'),
       );
-      this.#close(
-        Frame.protocolErrorClose(
-          PROTOCOL_ERROR_SUBTYPES.COMMON,
-          this.#isClient,
-        ),
-      );
+      this.#close(Frame.protocolErrorClose('COMMON', this.#isClient));
     }
   }
 
@@ -321,4 +299,4 @@ class Connection extends EventEmitter {
   }
 }
 
-module.exports = { Connection };
+module.exports = { Connection, CLOSE_TIMEOUT };
