+++ lib/websocket/frame.js
@@ -3,29 +3,91 @@
 const crypto = require('node:crypto');
 
 const {
-  FINAL_FRAME,
-  MASK_MASK,
-  LEN_16_BIT,
-  LEN_64_BIT,
   OPCODES,
-  MAX_16_BIT,
-  CONTROL_FRAME_MASK,
-  EMPTY_BUFFER,
-  RSV_MASK,
+  CLOSE_CODES,
   CONTROL_OPCODES,
   DATA_OPCODES,
-  RSV,
-  ENCODING,
-  CLIENT_EMPTY_PING,
-  EMPTY_PING,
-  CLIENT_EMPTY_PONG,
-  EMPTY_PONG,
-  CLOSE_FRAMES,
-  CLOSE_CODES,
-  MAX_CLOSE_REASON,
-  TWO_32,
 } = require('./constants.js');
 const { Result } = require('./result.js');
+
+const FINAL_FRAME = 0x80;
+const RSV = 0x00;
+const LEN_16_BIT = 126;
+const LEN_64_BIT = 127;
+const MAX_16_BIT = 65535;
+const TWO_32 = 4294967296;
+const MASK_MASK = 0x80;
+const RSV_MASK = 0x70;
+const CONTROL_FRAME_MASK = 0x08;
+const ENCODING = 'utf8';
+const MAX_CLOSE_REASON = 123;
+const EMPTY_BUFFER = Buffer.alloc(0);
+const EMPTY_PING = Buffer.from([0x89, 0x00]);
+const EMPTY_PONG = Buffer.from([0x8a, 0x00]);
+const CLIENT_EMPTY_PING = Buffer.from([0x89, 0x80, 0x00, 0x00, 0x00, 0x00]);
+const CLIENT_EMPTY_PONG = Buffer.from([0x8a, 0x80, 0x00, 0x00, 0x00, 0x00]);
+
+const CLOSE_FRAMES = {
+  NORMAL_CLOSE: {
+    reason: '',
+    buf: Buffer.from([0x88, 0x02, 0x03, 0xe8]),
+  },
+  GOING_AWAY: {
+    reason: 'Going away',
+    buf: Buffer.concat([
+      Buffer.from([0x88, 0x0c, 0x03, 0xe9]),
+      Buffer.from('Going away', ENCODING),
+    ]),
+  },
+  PROTOCOL_ERROR: {
+    COMMON: {
+      reason: 'Protocol error',
+      buf: Buffer.concat([
+        Buffer.from([0x88, 0x10, 0x3, 0xea]),
+        Buffer.from('Protocol error', ENCODING),
+      ]),
+    },
+    UNMASKED: {
+      reason: 'Unmasked frame from client',
+      buf: Buffer.concat([
+        Buffer.from([0x88, 0x1c, 0x03, 0xea]),
+        Buffer.from('Unmasked frame from client', ENCODING),
+      ]),
+    },
+    MASKED: {
+      reason: 'Masked frame from server',
+      buf: null,
+    },
+    RSV: {
+      reason: 'RSV bits must be 0',
+      buf: Buffer.concat([
+        Buffer.from([0x88, 0x14, 0x03, 0xea]),
+        Buffer.from('RSV bits must be 0', ENCODING),
+      ]),
+    },
+    CTRL_TOO_LONG: {
+      reason: 'Control frame too long',
+      buf: Buffer.concat([
+        Buffer.from([0x88, 0x18, 0x03, 0xea]),
+        Buffer.from('Control frame too long', ENCODING),
+      ]),
+    },
+  },
+  INVALID_PAYLOAD: {
+    reason: 'Invalid payload data',
+    buf: Buffer.concat([
+      Buffer.from([0x88, 0x16, 0x03, 0xef]),
+      Buffer.from('Invalid payload data', ENCODING),
+    ]),
+  },
+  MESSAGE_TOO_BIG: {
+    reason: 'Message too big',
+    buf: Buffer.concat([
+      Buffer.from([0x88, 0x11, 0x03, 0xf1]),
+      Buffer.from('Message too big', ENCODING),
+    ]),
+  },
+};
 
 const truncateCloseReason = (reason) => {
   const bytes = Buffer.from(reason, ENCODING);
