+++ lib/websocket/frameParser.js
@@ -1,23 +1,35 @@
 'use strict';
 
 const {
-  OPCODE_MASK,
-  MASK_MASK,
-  PAYLOAD_LEN_MASK,
-  LEN_16_BIT,
-  LEN_64_BIT,
-  FIN_MASK,
-  RSV_MASK,
   CONTROL_OPCODES,
   DATA_OPCODES,
   OPCODES,
-  VALID_CLOSE_CODES,
-  VALID_USER_CLOSE_CODES,
-  MAX_SAFE_HIGH_MASK,
-  TWO_32,
+  CLOSE_CODES,
 } = require('./constants.js');
 const { Frame } = require('./frame.js');
 const { Result } = require('./result.js');
+
+const FIN_MASK = 0x80;
+const RSV_MASK = 0x70;
+const OPCODE_MASK = 0x0f;
+const MASK_MASK = 0x80;
+const PAYLOAD_LEN_MASK = 0x7f;
+const LEN_16_BIT = 126;
+const LEN_64_BIT = 127;
+const MAX_SAFE_HIGH_MASK = 0xffe00000;
+const TWO_32 = 4294967296;
+const VALID_USER_CLOSE_CODES = { MIN: 3000, MAX: 4999 };
+const VALID_CLOSE_CODES = new Set([
+  CLOSE_CODES.NORMAL_CLOSE,
+  CLOSE_CODES.GOING_AWAY,
+  CLOSE_CODES.PROTOCOL_ERROR,
+  CLOSE_CODES.UNSUPPORTED_DATA,
+  CLOSE_CODES.INVALID_PAYLOAD,
+  CLOSE_CODES.POLICY_VIOLATED,
+  CLOSE_CODES.MESSAGE_TOO_BIG,
+  CLOSE_CODES.MANDATORY_EXTENSION,
+  CLOSE_CODES.INTERNAL_SERVER_ERROR,
+]);
 
 const isValidUTF8 = (buf) => {
   let i = 0;
