+++ lib/websocket/server.js
@@ -4,7 +4,17 @@ const crypto = require('node:crypto');
 const { EventEmitter } = require('node:events');
 
 const { Connection } = require('./connection.js');
-const { MAGIC, EOL, EOL2, UPGRADE, PING_INTERVAL } = require('./constants.js');
+
+const MAGIC = '258EAFA5-E914-47DA-95CA-C5AB0DC85B11';
+const PING_INTERVAL = 10000;
+const EOL = '\r\n';
+const EOL2 = '\r\n\r\n';
+const UPGRADE = [
+  'HTTP/1.1 101 Switching Protocols',
+  'Upgrade: websocket',
+  'Connection: Upgrade',
+  'Sec-WebSocket-Accept: ',
+].join(EOL);
 
 const hasToken = (value, token) =>
   !!value && value.toLowerCase().includes(token);
@@ -179,4 +189,4 @@ class WebsocketServer extends EventEmitter {
   }
 }
 
-module.exports = { WebsocketServer };
+module.exports = { WebsocketServer, MAGIC };
