+++ lib/websocket/ws.js
@@ -1,8 +1,8 @@
 'use strict';
 
-const constants = require('./constants.js');
-const { WebsocketServer } = require('./server.js');
-const { Connection } = require('./connection.js');
+const { OPCODES, CLOSE_CODES } = require('./constants.js');
+const { WebsocketServer, MAGIC } = require('./server.js');
+const { Connection, CLOSE_TIMEOUT } = require('./connection.js');
 const { Frame } = require('./frame.js');
 const {
   FrameParser,
@@ -11,7 +11,10 @@ const {
 } = require('./frameParser.js');
 
 module.exports = {
-  ...constants,
+  OPCODES,
+  CLOSE_CODES,
+  CLOSE_TIMEOUT,
+  MAGIC,
   WebsocketServer,
   Connection,
   Frame,
